refactor: reuse orders on dashboard
