+++ client-app/shared/account/composables/index.ts
@@ -2,7 +2,7 @@ export { default as useUser } from "./useUser";
 export { default as useUserAddresses } from "./useUserAddresses";
 export { default as useUserCheckoutDefaults } from "./useUserCheckoutDefaults";
 export { default as useUserOrder } from "./useUserOrder";
-export { default as useUserOrders } from "./useUserOrders";
+export * from "./useUserOrders";
 export { default as useUserOrdersFilter } from "./useUserOrdersFilter";
 export { default as useUserQuotes } from "./useUserQuotes";
 export { default as useUserQuote } from "./useUserQuote";
